refactor: add chartstatistics component, remove commented & unnecessary code
