--- my-app/src/components/statistics/ChartTitle.tsx
@@ -1,14 +1,0 @@
-import Typography from '@mui/material/Typography';
-import { ReactNode } from 'react';
-
-interface TitleProps {
-  children?: ReactNode;
-}
-
-export default function ChartTitle(props: TitleProps) {
-  return (
-    <Typography component="h2" variant="h6" color="primary" gutterBottom>
-      {props.children}
-    </Typography>
-  );
-}
