--- my-app/src/components/statistics/stat.module.css
@@ -1,4 +1,0 @@
-.statCont {
-    background-color: rgba(4, 95, 156, 0.658);
-    
-}
